requirements.txt
amqp==2.1.4
appdirs==1.4.2
arrow==0.10.0
asn1crypto==0.22.0
beautifulsoup4==4.5.3
billiard==3.5.0.2
blinker==1.4
boto3==1.4.4
botocore==1.5.22
celery==4.0.2
cffi==1.10.0
click==6.7
cryptography==1.8.1
docutils==0.13.1
enum34==1.1.6
Flask==0.12
Flask-Mail==0.9.1
flickr-api==0.5
funcsigs==1.0.2
futures==3.0.5
gunicorn==19.7.0
idna==2.5
inflection==0.3.1
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.5
jmespath==0.9.1
kombu==4.0.2
lxml==3.7.3
MarkupSafe==0.23
mock==2.0.0
ndg-httpsclient==0.4.2
numpy==1.12.1
oauth==1.0.1
olefile==0.44
packaging==16.8
pbr==3.0.1
Pillow==4.0.0
protobuf==3.3.0
pyasn1==0.2.3
pycparser==2.17
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
python-dotenv==0.6.3
pytz==2016.10
ratelimit==1.2.0
redis==2.10.5
requests==2.13.0
rq==0.7.1
s3transfer==0.1.10
six==1.10.0
tensorflow==1.1.0
vine==1.1.3
Werkzeug==0.12.2
